# backend/openapi-schema.yaml
openapi: 3.0.3
info:
  title: Shava API
  version: 1.0.0
  description: 'Public REST API for the Shava project: places, reviews, ratings, users
    (JWT auth), articles, gamification.'
paths:
...
  /api/v1/reviews/{id}/helpful/:
    post:
      operationId: reviews_helpful_create
      description: Idempotent on repeat calls. Combine with ``DELETE`` to toggle.
        Returns the new aggregate ``helpful_count`` and the viewer's current vote
        state.
      summary: Cast a 'helpful' vote on a review
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - reviews
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewHelpfulResponse'
          description: ''
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewHelpfulResponse'
          description: ''
        '400':
          description: Cannot vote on your own review.
        '401':
          description: Authentication required.
        '404':
          description: Review not found.
    delete:
      operationId: reviews_helpful_destroy
      description: Idempotent — safe to call when the user has not voted. Returns
        the new aggregate ``helpful_count`` and ``voted=false``.
      summary: Remove a 'helpful' vote on a review
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - reviews
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewHelpfulResponse'
          description: ''
        '401':
          description: Authentication required.
        '404':
          description: Review not found.
...
components:
  schemas:
...
    ReviewHelpfulResponse:
      type: object
      properties:
        helpful_count:
          type: integer
        voted:
          type: boolean
      required:
      - helpful_count
      - voted
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT